코드 간결화

+++ app/src/main/java/com/example/calculrator/MultiplyOperation.kt
@@ -3,5 +3,5 @@ package com.example.calculrator
 import com.example.calculator.AbstractOperation
 
 class MultiplyOperation : AbstractOperation () {
-    override fun operate(num1: Double, num2: Double): Double = (num1 * num2).toDouble()
+    override fun operate(num1: Double, num2: Double): Double = (num1 * num2)
 }
